Removed unuseful comments and fully replaced Parsec implementation with Megaparsec implementation.
