+++ stack.yaml
@@ -3,6 +3,6 @@ resolver: lts-11.14
 
 packages:
 
-# nix:
-#   enable: true
-#   shell-file: shell.nix
+nix:
+  enable: true
+  shell-file: shell.nix
