+++ test/ParserTests.hs
@@ -1,6 +1,6 @@
 module ParserTests where
 
-import SIL.ParserMegaparsec
+import SIL.Parser
 import Test.Tasty
 import Test.Tasty.HUnit
 import qualified Data.Map as Map
@@ -15,86 +15,39 @@ tests = testGroup "Tests" [unitTests]
 unitTests = testGroup "Unit tests"
   [ testCase "test Pair 0" $ do
       res <- runTestPair testPair0
-      let Right resParsec' = Parsec.runParsecParser Parsec.parsePair testPair0
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testPair0ParsecResult @?= EQ
   ,testCase "test ITE 1" $ do
       res <- runTestITE testITE1
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseITE testITE2
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testITEParsecResult @?= EQ
   , testCase "test ITE 2" $ do
       res <- runTestITE testITE2
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseITE testITE2
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testITEParsecResult @?= EQ
   , testCase "test ITE 3" $ do
       res <- runTestITE testITE3
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseITE testITE3
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testITEParsecResult @?= EQ
     -- Maybe TODO: Maybe Fix
     -- Probably thinks "then" is part of an assignment
   , testCase "test ITE 4" $ do
       res <- runTestITE testITE4
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseITE testITE4
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testITEParsecResult @?= EQ
   , testCase "test ITE with Pair" $ do
       res <- runSILParser parseITE testITEwPair
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseITE testITEwPair
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testITEWPairParsecResult @?= EQ
   , testCase "test Complete Lambda with ITE Pair" $ do
       res <- runSILParser parseCompleteLambda testCompleteLambdawITEwPair
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseCompleteLambda testCompleteLambdawITEwPair
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testCompleteLambdawITEwPairParsecResult @?= EQ
   , testCase "test Lambda with ITE Pair" $ do
       res <- runSILParser parseLambda testLambdawITEwPair
-      let Right resParsec' = Parsec.runParsecParser Parsec.parseLambda testLambdawITEwPair
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
-  , testCase "test parse Prelude.sil" $ do
-      res <- runTestParsePrelude
-      preludeFile <- Strict.readFile "Prelude.sil"
-      let Right resParsec' = Parsec.parsePrelude preludeFile
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testLambdawITEwPairParsecResult @?= EQ
   , testCase "test parse assignment with Complete Lambda with ITE with Pair" $ do
       res <- runSILParser parseAssignment testParseAssignmentwCLwITEwPair1
-      let Right resParsec' =
-            Parsec.runParsecParser Parsec.parseAssignment testParseAssignmentwCLwITEwPair1
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testParseAssignmentwCLwITEwPair1ParsecResult @?= EQ
   , testCase "test parseTopLevel with Complete Lambda with ITE with Pair" $ do
       res <- runSILParser parseTopLevel testParseTopLevelwCLwITEwPair
-      let Right resParsec' =
-            Parsec.runParsecParser Parsec.parseTopLevel testParseTopLevelwCLwITEwPair
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
+      res `compare` testParseTopLevelwCLwITEwPairParsecResult @?= EQ
   , testCase "test parseMain with CL with ITE with Pair" $ do
       res <- runTestMainwCLwITEwPair
-      preludeFile <- Strict.readFile "Prelude.sil"
-      let Right resParsec' = Parsec.parseMain Map.empty testMainwCLwITEwPair
-          resParsec = show resParsec'
-      res `compare` resParsec @?= EQ
-  , testCase "test Megaparsec vs Parsec implementation with tictactoe.sil" $ do
-      preludeFile <- Strict.readFile "Prelude.sil"
-      tictactoe <- Strict.readFile "tictactoe.sil"
-      let megaPrelude = case parsePrelude preludeFile of
-                          Right p -> p
-                          Left pe -> error . getErrorString $ pe
-          parsecPrelude = case Parsec.parsePrelude preludeFile of
-                            Right p -> p
-                            Left pe -> error . show $ pe
-          megaparsecRes = case parseMain megaPrelude $ tictactoe of
-                            Right x -> show x
-                            Left err -> getErrorString err
-          parsecRes = case Parsec.parseMain parsecPrelude $ tictactoe of
-                        Right x -> show x
-                        Left err -> show err
-      megaparsecRes `compare` parsecRes @?= EQ
+      res `compare` runTestMainwCLwITEwPairParsecResult @?= EQ
   ]
 
 runTestPair :: String -> IO String
@@ -119,7 +72,6 @@ testITE1 = unlines $
   , "else"
   , "  2"
   ]
-
 testITE2 = unlines $
   [ "if 1"
   , "  then"
@@ -255,6 +207,9 @@ runTestMainwCLwITEwPair = do
       Left err -> getErrorString err
   return res
 
+-- |Usefull to see if tictactoe.sil was correctly parsed
+-- and was usefull to compare with the deprecated SIL.Parser
+-- Parsec implementation
 testWtictactoe = do
   preludeFile <- Strict.readFile "Prelude.sil"
   tictactoe <- Strict.readFile "tictactoe.sil"
@@ -267,26 +222,19 @@ testWtictactoe = do
       Left err -> getErrorString err
   return res
 
+testPair0ParsecResult = "TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero)))))))))))) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero)"
 
--- unitTest2' parse s r = it s $ case parse s of
---   Left e -> expectationFailure $ concat ["failed to parse ", s, " ", show e]
---   Right g -> fmap (show . PrettyIExpr) (testEval g) >>= \r2 -> if r2 == r
---     then pure ()
---     else expectationFailure $ concat [s, " result ", r2]
+testITEParsecResult = "TITE (TPair TZero TZero) (TPair TZero TZero) (TPair (TPair TZero TZero) TZero)"
 
--- unitTests parse = do
---   let unitTestType = unitTestType' parse
---       unitTest2 = unitTest2' parse
---   describe "type checker" $ do
---     undefined
---   describe undefined undefined
+testITEWPairParsecResult = "TITE (TPair TZero TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))))) (TPair TZero TZero))"
 
--- main = do
---   preludeFile <- Strict.readFile "Prelude.sil"
---   let prelude = case parsePrelude preludeFile of
---         Right p -> p
---         Left pe -> error $ show pe
---       parse = parseMain prelude
---   hspec $ do
---     unitTests parse
---     --nexprTests
+testCompleteLambdawITEwPairParsecResult = "TNamedCompleteLam \"input\" (TITE (TPair TZero TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))))) (TPair TZero TZero)))"
+
+
+testLambdawITEwPairParsecResult = "TNamedLam \"input\" (TITE (TPair TZero TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))))) (TPair TZero TZero)))"
+
+runTestMainwCLwITEwPairParsecResult = "Pair (Defer (SetEnv (Pair (Gate (Pair Zero Zero)) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero))))))))))))))) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair (Pair Zero Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero) Zero))))))))))))) Zero))))) Zero"
+
+testParseAssignmentwCLwITEwPair1ParsecResult = "()"
+
+testParseTopLevelwCLwITEwPairParsecResult = "fromList [(\"main\",TNamedCompleteLam \"input\" (TITE (TPair TZero TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair (TPair TZero TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero) TZero))))))))))))))) TZero)))]"
